Clear trunk high-severity findings in dependabot, Attio attributes, and telemetry test fixture (#244)
## Why
`trunk check --all` surfaced four high-severity findings that were left
unaddressed: a missing Dependabot cooldown period (supply-chain risk
from newly published packages), a duplicate
`is_select_attribute_writable` definition in `libs/attio/attributes.py`
that silently shadowed the correct implementation, an autouse pytest
fixture flagged as unused, and a `force_flush` override on
`_NoNetworkLogExporter` whose parameter type didn't match its base
class.

## Trade-offs
`attributes.py` had two functions named `is_select_attribute_writable` —
one checking only `is_archived`, the other also checking `attribute_type
== "select"`. Kept the second (stricter, test-covered) definition and
removed the first rather than merging them, since no other code
referenced the deleted one and the existing test suite
(`tests/libs/attio/test_attributes.py`) already exercises the surviving
behavior including the non-select-type case.

## Risks
Low. The `attributes.py` change only removes dead code shadowed by a
later redefinition — confirmed via `uv run pytest
tests/libs/attio/test_attributes.py` (all 4
`is_select_attribute_writable` tests pass). The `conftest.py` type fix
only affects test-only telemetry mocking, not production code.

+++ tests/conftest.py
@@ -18,7 +18,7 @@ from libs.attio.sdk_boundary import get_attio_sdk_client_class  # noqa: E402
 
 
 @pytest.fixture(scope="session", autouse=True)
-def _neuter_telemetry_network() -> Iterator[None]:
+def _neuter_telemetry_network() -> Iterator[None]:  # pyright: ignore[reportUnusedFunction]
     """Stop telemetry unit tests from phoning home at interpreter shutdown.
 
     The telemetry tests (``tests/test_smoke_telemetry.py``,
@@ -73,7 +73,7 @@ def _neuter_telemetry_network() -> Iterator[None]:
         def export(self, batch: Any) -> Any:  # noqa: ARG002
             return LogRecordExportResult.SUCCESS
 
-        def force_flush(self, timeout_millis: int = 30000) -> bool:  # noqa: ARG002
+        def force_flush(self, timeout_millis: float = 30000) -> bool:  # noqa: ARG002
             return True
 
         def shutdown(self) -> None:
